Split ptrace.cc up into OS-dependent and OS-independent files
